% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-info.R
\name{AddBaseChangeInfo}
\alias{AddBaseChangeInfo}
\title{Calculates base change for RDD comparisons of JACUSA output.}
\usage{
AddBaseChangeInfo(l)
}
\arguments{
\item{l}{List object created by \code{Read()}.}
}
\value{
Returns a list of base calls with the additional baseChange fields.
}
\description{
\code{AddBaseChangeInfo} calculates base change for gDNA vs. cDNA comparisons and adds the 
result to the initial list object.
}
\examples{
data <- AddBaseChangeInfo(untr_hek293_rdds)
## plot distribution of base changes
barplot(Table(data))

}

