% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-info.R
\name{AddBaseInfo}
\alias{AddBaseInfo}
\title{Calculates a base vector for each sample of JACUSA output.}
\usage{
AddBaseInfo(l, collapse = FALSE)
}
\arguments{
\item{l}{List object created by \code{Read()}.}

\item{collapse}{Logical indicates if read counts of sample 2 should be merged - replicates are collapsed.}
}
\value{
Returns a list of base calls with additional base1 and base2 fields.
}
\description{
\code{AddBaseInfo} calculates a base vector for each sample of JACUSA output and adds the 
result to the initial list object.
}
\examples{
data <- AddBaseInfo(untr_hek293_rdds)
## plot distribution of bases in sample1
barplot(table(data$base1))

}

