% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{AddEditingFreqInfo}
\alias{AddEditingFreqInfo}
\title{Calculates editing frequency for RDDs in JACUSA output.}
\usage{
AddEditingFreqInfo(l)
}
\arguments{
\item{l}{List object created by \code{Read()}.}
}
\value{
Returns a list of base calls with the additional editingFreq fields.
}
\description{
\code{AddEditingFreqInfo} calculates the editing frequency for each replicate and an average 
for gDNA vs. cDNA comparisons. The result is added to the initial list object.
}
\examples{
## Read JACUSA result file hek293_untreated.out
data <- Read("hek293_untreated.out")
## AddEditingFreqInfo implicitly adds the baseChange field
data <- AddEditingFreqInfo(data)
## plot a boxplot of editing frequencies for each base change
boxplot(tapply(data$editingFreq, data$baseChange, c))

}

