% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{Coverage}
\alias{Coverage}
\title{Helper function}
\usage{
Coverage(sample, collapse = FALSE)
}
\arguments{
\item{sample}{todo}

\item{collapse}{todo}
}
\value{
todo
}
\description{
Calculate read coverage
}

