% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{FalsePositives}
\alias{FalsePositives}
\title{Helper function}
\usage{
FalsePositives(l, editing)
}
\arguments{
\item{l}{todo}

\item{editing}{todo}
}
\value{

}
\description{
assuming some true editing gives the fraction of false editing
}

