% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{FilterByCoverageHelper}
\alias{FilterByCoverageHelper}
\title{Helper function}
\usage{
FilterByCoverageHelper(covs, cov)
}
\arguments{
\item{covs}{todo}

\item{cov}{todo}
}
\value{
todo
}
\description{
FilterByCoverage Helper
}

