% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{FilterByMinVariantCount}
\alias{FilterByMinVariantCount}
\title{Filters JACUSA results of RDD calls and ensures minimal number of variant reads.}
\usage{
FilterByMinVariantCount(l, min_count = 2, collapse = F)
}
\arguments{
\item{l}{List object created by \code{Read()}.}

\item{min_count}{Numeric value that specifies the minimal number of variant reads in sample 2 or cDNA.}

\item{collapse}{Logical indicates if read counts of sample 2 should be merged - replicates are collapsed.}
}
\value{
Returns List of sites that have at least min_count variant reads in cDNA.
}
\description{
\code{FilterByMinVariant} Filters JACUSA result files of gDNA vs. cDNA comparisons and enforces a minimal number of variant bases in cDNA.
Per default gDNA is expected to be sample 1 and cDNA sample2!
}
\examples{
## Read JACUSA result file hek293_untreated.out
data <- Read("hek293_untreated.out")
## Filters sites that have less than 2 variant reads in the cDNA sample. 
data <- FilterByMinVariantCount(data, min_count = 2)

}

