% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{FilterRec}
\alias{FilterRec}
\title{Filters List of sites recursively.}
\usage{
FilterRec(l, f)
}
\arguments{
\item{l}{List object created by \code{Read()}.}

\item{f}{Vector of sites that should be retained.}
}
\value{
Returns List with sites that are contained in vector f.
}
\description{
\code{FilterRec} filters list of sites recursively.
}
\examples{
## calculate variant count
variant_count <- GetVariantCount(untr_hek293_rdds, collapse = TRUE)
## create index of sites that contain at least 10 variant bases
index <- variant_count >= 10
## filter data according to index
filtered <- FilterRec(untr_hek293_rdds, index)

}

