% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{GetVariantCount}
\alias{GetVariantCount}
\title{Calculates the number of variant reads in RDD JACUSA results files}
\usage{
GetVariantCount(l, collapse = F)
}
\arguments{
\item{l}{List object created by \code{Read()}.}

\item{collapse}{Logical indicates if read counts of sample 2 should be merged - replicates are collapsed.}
}
\value{
Returns List of count that represent the number of variant reads in cDNA.
}
\description{
Calculates the number of variant reads in JACUSA results of RDD calls
}
\details{
Calculates the number of variant reads in JACUSA result files of gDNA vs. cDNA comparisons.
Per default gDNA is expected to be sample 1 and cDNA to be sample 2!
}
\examples{
## Read JACUSA result file JACUSA.out
data <- Read(data)
variant_count <- GetVariantCount(data)
}

