% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{PlotTable}
\alias{PlotTable}
\title{Helper function}
\usage{
PlotTable(tbl, score = T)
}
\arguments{
\item{tbl}{todo}

\item{score}{todo}
}
\value{

}
\description{
plots the distribution of base changes
}

