% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{Samples}
\alias{Samples}
\title{Returns base call columns for a sample from a JACUSA result file.}
\usage{
Samples(l, sample)
}
\arguments{
\item{l}{List object created by \code{Read()}.}

\item{sample}{Integer value: 1 or 2.}
}
\value{
Returns a list of base calls "," separated for choosen sample.
}
\description{
\code{Samples} returns base call columns for a sample (1 or 2) from a JACUSA result file.
}
\examples{
## Read JACUSA result file hek293_untreated.out
data <- Read("hek293_untreated.out")
## Extract sequencing information for sample 1. 
sample1 <- Samples(data, 1)

}

