% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{Score}
\alias{Score}
\title{Calculates the fraction of editing sites among RDDs in JACUSA output.}
\usage{
Score(tbl, editing = c("A->G"))
}
\arguments{
\item{tbl}{Vector object created by \code{Table()}.}

\item{editing}{Vector of strings that identifies true editing, e.g.: "A->G".}
}
\value{
Returns a numeric values that represent the fraction of editing sites among RDDs.
}
\description{
\code{Score} calculates the fraction of editing sites among RDDs in JACUSA output.
}
\examples{
## Read JACUSA result file hek293_untreated.out
data <- Read("hek293_untreated.out")
## Table implicitly populates the baseChange field
tbl <- Table(data)
Score(tbl)

}

