% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ToMatrix}
\alias{ToMatrix}
\title{Converts base counts from sample columns to matrices.}
\usage{
ToMatrix(sample, invert = F, collapse = T)
}
\arguments{
\item{sample}{List of string vectors where base counts (A,C,G,T) are separated by ",". Typically, the result of \code{Samples(data, i)}, where i = 1 or 2.}

\item{invert}{Logical indicates if base counts should be inverted (A=>T, C=>G, etc...) after conversion. Depends on employed sequencing library.}

\item{collapse}{Logical indicates if sample is a list of string vectors and if converted matrices should be aggregated - collapses replicates.}
}
\value{
Returns a matrix or a list of matrices of base counts.
}
\description{
\code{ToMatrix} converts base counts encoded as "," separated string vectors to count matrices.
}
\examples{
## Read JACUSA result file hek293_untreated.out asdf
data <- Read("hek293_untreated.out")
## Extract sequencing info of sample 1
sample1 <- Samples(data, 1)
## Convert character encoded base counts to count matrices
matrix1 <- ToMatrix(sample1)

}

