% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{untr_hek293_rdds}
\alias{untr_hek293_rdds}
\title{Subset of RDDs detected by JACUSA in HEK-293 untreated cells}
\format{a list with 11 elements where each has 10,000 elements on its on.
\itemize{
	\item contig: contig
	\item start: position of variant (>=0)
	\item end: start + 1 (conform to BED file format)
	\item name: arbitrary (conform to BED file format)
	\item stat: higher value of test-statistic indicate more divergent pileups)
	\item strand: "+", "-", "." (no strand available)
	\item bases11: Base count for genomic DNA (A, C, G, T)
	\item bases21: Base count for RNA replicate 1
	\item bases22: Base count for RNA replicate 2
	\item info: Additional info for this specific site. Empty field is equal to "*"
	\item filter_info: comma-separated list of feature filters Empty field is equal equal "*"
}}
\usage{
untr_hek293_rdds
}
\description{
A dataset containing a subset of RNA DNA differences (RDDs) identified by JACUSA in untreated HEK-293 cells. 
The fields are as follows:
}
\keyword{datasets}

