% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Write}
\alias{Write}
\title{Writes a list of sites to a file.}
\usage{
Write(l, file, extra = NULL)
}
\arguments{
\item{l}{Vector object created by \code{Table()}.}

\item{file}{String is the filename to store the list object.}

\item{extra}{Vector of strings that defines additional elements from the list that 
will be stored in the file.}
}
\description{
\code{Write} Stores a list of sites in a file.
}

