% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{AddBaseInfo}
\alias{AddBaseInfo}
\title{Calculates a base vector for each sample of JACUSA output.}
\usage{
AddBaseInfo(l)
}
\arguments{
\item{l}{List object created by \code{Read()}.}
}
\value{
Returns a list of base calls with additional base1 and base2 fields.
}
\description{
\code{AddBaseInfo} calculates a base vector for each sample of JACUSA output and adds the result to the initial list object.
}
\examples{
## Read JACUSA result file JACUSA.out
data <- Read(data)
data <- AddBaseInfo(data)
## plot distribution of bases in sample1
barplot(table(data$base1))
}

