% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{FilterResult}
\alias{FilterResult}
\title{Retains sites that contain the variant base in all replicates of at least one sample.}
\usage{
FilterResult(l)
}
\arguments{
\item{l}{List object created by \code{Read()}.}
}
\value{
Returns List with sites where at least one sample contains the variant base in all replicates.
}
\description{
\code{FilterResults} Enforces that at least one sample contains the variant base in all replicates.
}
\examples{
result <- FilterResult(untr_hek293_rdds)

}

