% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Read}
\alias{Read}
\title{Reads JACUSA output.}
\usage{
Read(f, invert = FALSE, stat = NULL, fields = NULL, cov = NULL,
  collapse = FALSE, ...)
}
\arguments{
\item{f}{String represents the filename of the JACUSA output.}

\item{invert}{Logical indicates if base calls should be inverted.}

\item{stat}{Numeric value represents the minimal test-statistic.}

\item{fields}{Vector of strings defines how to filter by read coverage}

\item{cov}{Vector of numeric defines the minimal read coverage.}

\item{collapse}{Logical indicates if replicates should be collapsed - replicates of base 
count vectors will be aggregated.}

\item{...}{Additional parameters that will be forwarded to read.table.}
}
\value{
Returns a list of base calls "," separated for chosen parameters.
}
\description{
\code{Read} reads JACUSA output and returns a list of identified sites. See JACUSA manual for 
details on encoding of replicates and samples.
}

