% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{Write}
\alias{Write}
\title{Writes a list of sites to a file.}
\usage{
Write(l, file, extra = NULL)
}
\arguments{
\item{l}{Vector object created by \code{Table()}.}

\item{file}{String is the filename to store the list object.}

\item{extra}{Vector of strings that defines additional elements from the list that will be stored in the file.}
}
\description{
\code{Write} Stores a list of sites in a file.
}
\examples{
## Read JACUSA result file JACUSA.out
data <- Read(data)
data <- AddBaseChangeInfo(data)
## base change will be stored in the id/name column according to the BED file format definition.
data$name <- data$baseChange
Write(data)
}

